// content/image-manifest.generated.ts
export const imageManifest = {
  "collage/collage-28699.jpg": {
    "source": "/images/collage/collage-28699.jpg",
    "width": 3900,
    "height": 4003,
    "dominantColor": "#767269",
    "blurDataUrl": "data:image/jpeg;base64,/9j/2wBDABELDA8MChEPDg8TEhEUGSobGRcXGTMkJh4qPDU/Pjs1OjlDS2BRQ0daSDk6U3FUWmNma2xrQFB2fnRofWBpa2f/2wBDARITExkWGTEbGzFnRTpFZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2f/wAARCAAQABADASIAAhEBAxEB/8QAFgABAQEAAAAAAAAAAAAAAAAABAED/8QAIhAAAgEEAQQDAAAAAAAAAAAAAQIEAxESIQAFIjJBcYGR/8QAFQEBAQAAAAAAAAAAAAAAAAAAAQL/xAAYEQADAQEAAAAAAAAAAAAAAAAAAQIxYf/aAAwDAQACEQMRAD8Ax6bGIiIAMWZNXJ0bcRVp0UgO0d38T3ObWOOz8cJAkpRgU8iSMDdl1b79ckydGerpMgRiquAcj7JH5vkw9Glw/9k="
  },
// ...
  "collage/collage-img-0953.jpg": {
    "source": "/images/collage/collage-img-0953.jpg",
    "width": 356,
    "height": 1206,
    "dominantColor": "#87766E",
    "blurDataUrl": "data:image/jpeg;base64,/9j/2wBDABELDA8MChEPDg8TEhEUGSobGRcXGTMkJh4qPDU/Pjs1OjlDS2BRQ0daSDk6U3FUWmNma2xrQFB2fnRofWBpa2f/2wBDARITExkWGTEbGzFnRTpFZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2f/wAARCAAQAAUDASIAAhEBAxEB/8QAFQABAQAAAAAAAAAAAAAAAAAAAAT/xAAfEAACAgIBBQAAAAAAAAAAAAABAgADBBESEyJhgcH/xAAUAQEAAAAAAAAAAAAAAAAAAAAD/8QAFxEBAQEBAAAAAAAAAAAAAAAAAQARAv/aAAwDAQACEQMRAD8AiycyrBfpPULLD3MSSNE+ol1j3qEUVtwA0pq0S3k/IhK7OdAZf//Z"
  },
  "collage/collage-img-3149.png": {
    "source": "/images/collage/collage-img-3149.png",
    "width": 1824,
    "height": 3464,
    "dominantColor": "#87624C",
    "blurDataUrl": "data:image/jpeg;base64,/9j/2wBDABELDA8MChEPDg8TEhEUGSobGRcXGTMkJh4qPDU/Pjs1OjlDS2BRQ0daSDk6U3FUWmNma2xrQFB2fnRofWBpa2f/2wBDARITExkWGTEbGzFnRTpFZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2f/wAARCAAQAAgDASIAAhEBAxEB/8QAFgABAQEAAAAAAAAAAAAAAAAAAAME/8QAIBAAAQQCAQUAAAAAAAAAAAAAAQACAxESITEEEzNBgf/EABUBAQEAAAAAAAAAAAAAAAAAAAME/8QAFREBAQAAAAAAAAAAAAAAAAAAEQD/2gAMAwEAAhEDEQA/AMc8DhAOob43VWjk0HVj6EVHPMUPbkLrxJGDdc+q4O0QDUpf/9k="
  },
  "collage/collage-img-4899.png": {
    "source": "/images/collage/collage-img-4899.png",
    "width": 1360,
    "height": 3737,
    "dominantColor": "#685B50",
    "blurDataUrl": "data:image/jpeg;base64,/9j/2wBDABELDA8MChEPDg8TEhEUGSobGRcXGTMkJh4qPDU/Pjs1OjlDS2BRQ0daSDk6U3FUWmNma2xrQFB2fnRofWBpa2f/2wBDARITExkWGTEbGzFnRTpFZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2f/wAARCAAQAAYDASIAAhEBAxEB/8QAFgABAQEAAAAAAAAAAAAAAAAAAAQF/8QAHxAAAgICAQUAAAAAAAAAAAAAAQIABAMRIQUSUZHw/8QAFQEBAQAAAAAAAAAAAAAAAAAAAgP/xAAXEQADAQAAAAAAAAAAAAAAAAAAAREh/9oADAMBAAIRAxEAPwDOx1ciFWdSFK8dvPuJV0y8tW9aYoxw5DtPI+3EniHWf//Z"
  },
// ...
  "work-05-self-collage.jpg": {
    "source": "/images/work-05-self-collage.jpg",
    "width": 5057,
    "height": 7150,
    "dominantColor": "#8A837F",
    "blurDataUrl": "data:image/jpeg;base64,/9j/2wBDABELDA8MChEPDg8TEhEUGSobGRcXGTMkJh4qPDU/Pjs1OjlDS2BRQ0daSDk6U3FUWmNma2xrQFB2fnRofWBpa2f/2wBDARITExkWGTEbGzFnRTpFZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2dnZ2f/wAARCAAQAAsDASIAAhEBAxEB/8QAFQABAQAAAAAAAAAAAAAAAAAABAX/xAAjEAACAQQBAwUAAAAAAAAAAAABAgMABBEhEiIxQRNCYbHw/8QAFAEBAAAAAAAAAAAAAAAAAAAAAf/EABYRAQEBAAAAAAAAAAAAAAAAAAARQf/aAAwDAQACEQMRAD8ARbusKsAhjmKg4U5CjWqky2t9LK0h9xz1Lk/VVLa8jSykiE3Fk2MjkfGR+7UYKT1KyyA75euV7/HijTH/2Q=="
  }
} satisfies Readonly<Record<string, ImageManifestEntry>>;